maskchat/mask-chat/requirements.txt
Flask==3.0.0
python-socketio==5.8.0
eventlet==0.33.0